Fix null pointer dereferences in obs_backscrub_create
In this function, if bs_maskgen_new fails, filter will be set to NULL.
The very next line tries to write into other fields of filter, causing
a segfault. To avoid this, return early from the function. Also adds a
check to avoid even trying to call into libbackscrub if we don't have
a valid modelname.